Modularice aurora, add png circuit
